fix: create utils function separately
