+++ src/pages/Assignments/index.tsx
@@ -1,6 +1,7 @@
 import { useState, type JSX, useEffect } from 'react'
 import ASSIGNMENTS from '../../assets/mockData/assignments'
-import AssessmentCard from './AssessmentCard'
+import AssessmentCard from './AssignmentCard'
+import { filterAssignments, getTabButtons } from '../../utils/assignmentsUtils'
 
 const AssignmentPage = (): JSX.Element => {
   const [selectedCourse, setSelectedCourse] = useState(ASSIGNMENTS[0].course)
@@ -87,43 +88,3 @@ const AssignmentPage = (): JSX.Element => {
 }
 export default AssignmentPage
 
-const getTabButtons = (
-  assignments: (typeof ASSIGNMENTS)[number]['assignments']
-): string[] => {
-  const pendingAssignments = assignments.filter(
-    (assignment) => assignment.status === 'pending'
-  )
-  const completedAssignments = assignments.filter(
-    (assignment) => assignment.status === 'completed'
-  )
-  const submittedAssignments = assignments.filter(
-    (assignment) => assignment.status === 'submitted'
-  )
-  const tabButtons = [
-    'All',
-    `Pending (${pendingAssignments?.length})`,
-    `Submitted (${submittedAssignments?.length})`,
-    `Completed (${completedAssignments?.length})`,
-  ]
-  return tabButtons
-}
-
-const filterAssignments = (
-  courseData: (typeof ASSIGNMENTS)[number],
-  selectedAssignmentType: string
-): Array<{
-  status: 'submitted' | 'pending' | 'completed'
-  title: string
-  dueDate: string
-  tasks: Array<{
-    title: string
-    status: 'submitted' | 'pending' | 'checked'
-  }>
-}> => {
-  const data = courseData.assignments.filter(
-    (assignment) =>
-      assignment.status === selectedAssignmentType.split(' ')[0].toLowerCase()
-  )
-  if (data.length === 0) return courseData.assignments
-  return data
-}
